Allow LocParams to run on non-Full parameters in Menu.
We now continue the menu processing pipeline even if the
Menu parameter(s) is/are non-Full. When we get to the point
of rendering the template, if there is no rendered early
response and the Menu parameter is non-Full, we render
the usual 404. If there is a value, we do the usual template
processing.

As such, everything except for template processing will run
even if the request parameter could not be deserialized
properly. Note that the request path will still need to match.