First pass at a c"" macro for CssSelectors.
Right now it does a compile-time parse of the selectors and
emits literal constructors for the selector objects into the
code.

The next step is to do some additional analysis of the
resulting selectors and wrap selectors with indications of
what they’re modifying at their end. That’ll let us be smarter
about what CanBinds can receive or not from a particular
selector.