The c macro now infers AttrModifyingSelector.
The c macro can now be passed a regular CSS selector
and, if the parsed selector will modify an attribute, wraps the
result in an AttrModifyingSelector with the appropriate info.

This will allow binding functions that will be able to take the
attribute that will be bound as one of their parameters.