First pass at adding (better) search to API docs.
Right now we mostly just have the API docs builder script
run through a documentation helper that adds a JS file to
HTML files.

For now, we’re only doing the HTML files at the root of the
documentation directory, which isn’t terribly useful, but it does
work. More to come.