Add ParseCommonMark app for use with stmd spec runner.
This is an app, built in the test scope, that is designed to be used in
conjunction with the stmd spec runner to verify that Lift's CommonMark password
adheres to the CommonMark spec. Right now it's designed to be run from sbt with
output redirection (it prints to stderr, so that in a shell script it can be
invoked with sbt's output redirected to /dev/null and ParseCommonMark's output
redirected to stdout).